+++ src/WebJobs.Script.Rpc/Http/Configuration/CustomHandlerType.cs
