+++ src/WebJobs.Script.Rpc/Http/Configuration/HttpWorkerOptions.cs
