+++ src/WebJobs.Script.Rpc/Http/Configuration/HttpWorkerOptionsSetup.cs
