+++ src/WebJobs.Script.Rpc/Http/CustomHandlerHttpOptions.cs
