+++ src/WebJobs.Script.Rpc/Http/DefaultHttpWorkerService.cs
