+++ src/WebJobs.Script.Rpc/Http/HttpScriptInvocationContext.cs
