+++ src/WebJobs.Script.Rpc/Http/HttpScriptInvocationResult.cs
