+++ src/WebJobs.Script.Rpc/Http/HttpScriptInvocationResultExtensions.cs
