+++ src/WebJobs.Script.Rpc/Http/HttpWorkerChannel.cs
