+++ src/WebJobs.Script.Rpc/Http/HttpWorkerChannelFactory.cs
