+++ src/WebJobs.Script.Rpc/Http/HttpWorkerConstants.cs
