+++ src/WebJobs.Script.Rpc/Http/HttpWorkerContext.cs
