+++ src/WebJobs.Script.Rpc/Http/HttpWorkerDescription.cs
