+++ src/WebJobs.Script.Rpc/Http/HttpWorkerFunctionProvider.cs
