+++ src/WebJobs.Script.Rpc/Http/HttpWorkerProcess.cs
