+++ src/WebJobs.Script.Rpc/Http/HttpWorkerProcessFactory.cs
