+++ src/WebJobs.Script.Rpc/Http/HttpWorkerRoute.cs
