+++ src/WebJobs.Script.Rpc/Http/IHttpWorkerChannel.cs
