+++ src/WebJobs.Script.Rpc/Http/IHttpWorkerChannelFactory.cs
