+++ src/WebJobs.Script.Rpc/Http/IHttpWorkerProcessFactory.cs
