+++ src/WebJobs.Script.Rpc/Http/IHttpWorkerService.cs
